refactor: Wave 4 quality push - SRP, DRY, orthogonality (#1514, #1515, #1518)
Split contracts.py (868 lines) into focused package with 6 sub-modules:
- contracts/level.py: enforcement level configuration
- contracts/exceptions.py: exception hierarchy
- contracts/primitives.py: require()/ensure() functions
- contracts/decorators.py: @precondition, @postcondition, etc.
- contracts/validators.py: numeric/array validation helpers
- contracts/invariants.py: ContractChecker mixin and @invariant
- contracts/__init__.py: re-exports everything for backward compat

Decompose golf_gui_application.py (1663 lines) into 4 focused modules:
- golf_playback_controller.py: SmoothPlaybackController
- golf_gui_tabs.py: MotionCaptureTab, SimulinkModelTab, ComparisonTab
- golf_visualizer_widget.py: GolfVisualizerWidget (OpenGL)
- golf_gui_styles.py: QSS stylesheet constants

Centralize os.environ.get calls through config/environment.py:
- Add typed accessors: get_golf_port, get_golf_suite_mode, is_auth_disabled,
  get_golf_ui_dist, is_browser_suppressed, is_headless, get_display, get_dbc_level
- Migrate api/local_server.py, api/auth/middleware.py, api/diagnostics.py,
  shared/python/ui/adapters/canvas.py to use centralized accessors

Closes #1514, closes #1515, closes #1518
